Safari-609.1.11.1/LayoutTests/svg/animations/animations-paused-in-background-page-expected.txt

Tests that animated SVG images are properly paused when the page is no longer visible

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isImageAnimating(testImage) is true
Setting page visibility to hidden
PASS internals.isImageAnimating(testImage) became false
Setting page visibility to visible
PASS internals.isImageAnimating(testImage) became true
PASS successfullyParsed is true

TEST COMPLETE
